+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning].
 ## [Unreleased]
 ### Added
 - `add_transaction` method
+- `get_invoice` method
 
 ### Changed
 - Removed `offset` param for paginated calls, use `limitstart` instead.